requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cycler==0.10.0
fastapi==0.68.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.19.1
idna==3.2
jws==0.1.3
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
oauth2client==4.1.3
Pillow==8.0.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.4
pydantic==1.8.2
pyparsing==2.4.7
Pyrebase4==4.5.0
python-dateutil==2.8.1
python-jwt==2.0.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
itsdangerous
httpx
authlib
